+++ src/components/Header/hamburger.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import theme from '../styles/theme';
+import theme from '../../styles/theme';
 import styled from 'styled-components';
-import preventScroll from '../utils/preventScroll';
+import preventScroll from '../../utils/preventScroll';
 
 const MenuContainer = styled.div`
   width: 25%;
