+++ src/components/Header/index.ts
@@ -1,0 +1,2 @@
+export { default as Header } from './header';
+export { default as Hamburger } from './hamburger';
